[#84] Chart 개발 - DataStore 구조 개선 / Sunburst Legend, Resizing ##########  - DataStore 구조 개선    - 중복코드 제거를 위한 구조 변경  - Sunburst Legend 처리 및 Resizing 처리
