+++ examples/routers/charts/chart.line.vue
@@ -6,11 +6,11 @@
     />
     <chart
       :data="fillData"
-      :options="lineOptions"
+      :options="fillOptions"
     />
     <chart
       :data="stackData"
-      :options="lineOptions"
+      :options="stackOptions"
     />
     <chart
       :data="scatterData"
@@ -33,12 +33,36 @@
             {
               id: 'simple1',
               name: 'simple1',
-              point: false,
+              data: [
+                { x: '2018-05-25 05:11:00', y: 20 },
+                { x: '2018-05-25 05:12:00', y: 10 },
+                { x: '2018-05-25 05:13:00', y: 10 },
+                { x: '2018-05-25 05:14:00', y: 13 },
+                { x: '2018-05-25 05:15:00', y: 20 },
+                { x: '2018-05-25 05:16:00', y: 20 },
+                { x: '2018-05-25 05:17:00', y: 10 },
+                { x: '2018-05-25 05:18:00', y: 10 },
+                { x: '2018-05-25 05:19:00', y: 13 },
+                { x: '2018-05-25 05:20:00', y: 20 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
             {
               id: 'simple2',
               name: 'simple2',
-              point: false,
+              data: [
+                { x: '2018-05-25 05:11:00', y: 15 },
+                { x: '2018-05-25 05:12:00', y: 17 },
+                { x: '2018-05-25 05:13:00', y: 20 },
+                { x: '2018-05-25 05:14:00', y: 30 },
+                { x: '2018-05-25 05:15:00', y: 20 },
+                { x: '2018-05-25 05:16:00', y: 10 },
+                { x: '2018-05-25 05:17:00', y: 25 },
+                { x: '2018-05-25 05:18:00', y: 31 },
+                { x: '2018-05-25 05:19:00', y: 11 },
+                { x: '2018-05-25 05:20:00', y: 17 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
           ],
         },
@@ -47,14 +71,36 @@
             {
               id: 'simple1',
               name: 'simple1',
-              point: true,
-              fill: true,
+              data: [
+                { x: '2018-05-25 05:11:00', y: 20 },
+                { x: '2018-05-25 05:12:00', y: 10 },
+                { x: '2018-05-25 05:13:00', y: 10 },
+                { x: '2018-05-25 05:14:00', y: 13 },
+                { x: '2018-05-25 05:15:00', y: 20 },
+                { x: '2018-05-25 05:16:00', y: 20 },
+                { x: '2018-05-25 05:17:00', y: 10 },
+                { x: '2018-05-25 05:18:00', y: 10 },
+                { x: '2018-05-25 05:19:00', y: 13 },
+                { x: '2018-05-25 05:20:00', y: 20 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
             {
               id: 'simple2',
               name: 'simple2',
-              point: true,
-              fill: true,
+              data: [
+                { x: '2018-05-25 05:11:00', y: 15 },
+                { x: '2018-05-25 05:12:00', y: 17 },
+                { x: '2018-05-25 05:13:00', y: 20 },
+                { x: '2018-05-25 05:14:00', y: 30 },
+                { x: '2018-05-25 05:15:00', y: 20 },
+                { x: '2018-05-25 05:16:00', y: 10 },
+                { x: '2018-05-25 05:17:00', y: 25 },
+                { x: '2018-05-25 05:18:00', y: 31 },
+                { x: '2018-05-25 05:19:00', y: 11 },
+                { x: '2018-05-25 05:20:00', y: 17 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
           ],
         },
@@ -64,22 +110,55 @@
               id: 'simple1',
               name: 'simple1',
               point: true,
-              fill: true,
-              stack: true,
+              data: [
+                { x: '2018-05-25 05:11:00', y: 10 },
+                { x: '2018-05-25 05:12:00', y: 10 },
+                { x: '2018-05-25 05:13:00', y: 10 },
+                { x: '2018-05-25 05:14:00', y: 13 },
+                { x: '2018-05-25 05:15:00', y: 20 },
+                { x: '2018-05-25 05:16:00', y: 20 },
+                { x: '2018-05-25 05:17:00', y: 10 },
+                { x: '2018-05-25 05:18:00', y: 10 },
+                { x: '2018-05-25 05:19:00', y: 13 },
+                { x: '2018-05-25 05:20:00', y: 20 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
             {
               id: 'simple2',
               name: 'simple2',
               point: true,
-              fill: true,
-              stack: true,
+              data: [
+                { x: '2018-05-25 05:11:00', y: 15 },
+                { x: '2018-05-25 05:12:00', y: 17 },
+                { x: '2018-05-25 05:13:00', y: 20 },
+                { x: '2018-05-25 05:14:00', y: 30 },
+                { x: '2018-05-25 05:15:00', y: 20 },
+                { x: '2018-05-25 05:16:00', y: 10 },
+                { x: '2018-05-25 05:17:00', y: 25 },
+                { x: '2018-05-25 05:18:00', y: 31 },
+                { x: '2018-05-25 05:19:00', y: 11 },
+                { x: '2018-05-25 05:20:00', y: 17 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
             {
               id: 'simple3',
               name: 'simple3',
               point: true,
-              fill: true,
-              stack: true,
+              data: [
+                { x: '2018-05-25 05:11:00', y: 12 },
+                { x: '2018-05-25 05:12:00', y: 13 },
+                { x: '2018-05-25 05:13:00', y: 5 },
+                { x: '2018-05-25 05:14:00', y: 7 },
+                { x: '2018-05-25 05:15:00', y: 14 },
+                { x: '2018-05-25 05:16:00', y: 5 },
+                { x: '2018-05-25 05:17:00', y: 15 },
+                { x: '2018-05-25 05:18:00', y: 11 },
+                { x: '2018-05-25 05:19:00', y: 7 },
+                { x: '2018-05-25 05:20:00', y: 10 },
+                { x: '2018-05-25 05:21:00', y: 8 },
+              ],
             },
           ],
         },
@@ -90,23 +169,45 @@
               name: 'simple1',
               point: true,
               pointStyle: 'crossRot',
+              data: [
+                { x: '2018-05-25 05:11:00', y: 10 },
+                { x: '2018-05-25 05:12:00', y: 10 },
+                { x: '2018-05-25 05:13:00', y: 10 },
+                { x: '2018-05-25 05:14:00', y: 13 },
+                { x: '2018-05-25 05:15:00', y: 20 },
+                { x: '2018-05-25 05:16:00', y: 20 },
+                { x: '2018-05-25 05:17:00', y: 10 },
+                { x: '2018-05-25 05:18:00', y: 10 },
+                { x: '2018-05-25 05:19:00', y: 13 },
+                { x: '2018-05-25 05:20:00', y: 20 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
             {
               id: 'simple2',
               name: 'simple2',
               point: true,
               pointStyle: 'triangle',
+              data: [
+                { x: '2018-05-25 05:11:00', y: 12 },
+                { x: '2018-05-25 05:12:00', y: 13 },
+                { x: '2018-05-25 05:13:00', y: 5 },
+                { x: '2018-05-25 05:14:00', y: 7 },
+                { x: '2018-05-25 05:15:00', y: 14 },
+                { x: '2018-05-25 05:16:00', y: 20 },
+                { x: '2018-05-25 05:17:00', y: 10 },
+                { x: '2018-05-25 05:18:00', y: 10 },
+                { x: '2018-05-25 05:19:00', y: 13 },
+                { x: '2018-05-25 05:20:00', y: 20 },
+                { x: '2018-05-25 05:21:00', y: 20 },
+              ],
             },
           ],
         },
         lineOptions: {
           type: 'line',
-          width: '800px',
+          width: '550px',
           height: '230px',
-          legend: {
-            show: true,
-            position: 'right',
-          },
           title: {
             text: 'Title Test',
             show: true,
@@ -115,20 +216,60 @@
             scaleType: 'fix', // auto, fix, step
             labelType: 'time', // time, linear, category
             tickFormat: 'HH:mm:ss',
-            showGrid: false,
-            position: 'bottom',
             interval: 'minute',
           }],
           yAxes: [{
             scaleType: 'auto', // auto, fix, step
             labelType: 'linear', // time, linear, category
-            showGrid: true,
+            position: 'left',
+          }],
+        },
+        fillOptions: {
+          type: 'line',
+          width: '550px',
+          height: '230px',
+          fill: true,
+          title: {
+            text: 'Title Test',
+            show: true,
+          },
+          xAxes: [{
+            scaleType: 'fix', // auto, fix, step
+            labelType: 'time', // time, linear, category
+            tickFormat: 'HH:mm:ss',
+            interval: 'minute',
+          }],
+          yAxes: [{
+            scaleType: 'auto', // auto, fix, step
+            labelType: 'linear', // time, linear, category
+            position: 'left',
+          }],
+        },
+        stackOptions: {
+          type: 'line',
+          width: '550px',
+          height: '230px',
+          fill: true,
+          stack: true,
+          title: {
+            text: 'Title Test',
+            show: true,
+          },
+          xAxes: [{
+            scaleType: 'fix', // auto, fix, step
+            labelType: 'time', // time, linear, category
+            tickFormat: 'HH:mm:ss',
+            interval: 'minute',
+          }],
+          yAxes: [{
+            scaleType: 'auto', // auto, fix, step
+            labelType: 'linear', // time, linear, category
             position: 'left',
           }],
         },
         scatterOptions: {
           type: 'scatter',
-          width: '800px',
+          width: '550px',
           height: '230px',
           title: {
             text: 'Title Test1',
@@ -150,7 +291,7 @@
             position: 'left',
           }],
         },
-        lastData: '2018-05-25 12:15:00',
+        lastData: '2018-05-25 05:21:00',
       };
     },
     mounted() {
@@ -176,20 +317,19 @@
 
         this.$children[0].addValue(0, { x: this.lastData, y: randomData1 });
         this.$children[0].addValue(1, { x: this.lastData, y: randomData2 });
+        this.$children[0].chart.redraw();
 
         this.$children[1].addValue(0, { x: this.lastData, y: randomData3 });
         this.$children[1].addValue(1, { x: this.lastData, y: randomData4 });
+        this.$children[1].chart.redraw();
 
         this.$children[2].addValue(0, { x: this.lastData, y: randomData5 });
         this.$children[2].addValue(1, { x: this.lastData, y: randomData6 });
         this.$children[2].addValue(2, { x: this.lastData, y: randomData7 });
+        this.$children[2].chart.redraw();
 
         this.$children[3].addValue(0, { x: this.lastData, y: randomData8 });
         this.$children[3].addValue(1, { x: this.lastData, y: randomData9 });
-
-        this.$children[0].chart.redraw();
-        this.$children[1].chart.redraw();
-        this.$children[2].chart.redraw();
         this.$children[3].chart.redraw();
       },
     },
