+++ src/components/chart/core/axis/axis.scale.steps.js
@@ -6,9 +6,9 @@ class AxisStepsScale extends Axis {
     // init variable
     const options = this.options;
     const chartRect = this.chartRect;
-    const dataSet = this.dataSet;
+    const dataStore = this.dataStore;
 
-    const maxLabelInfo = dataSet.getLabelTextMaxInfo();
+    const maxLabelInfo = dataStore.getLabelTextMaxInfo();
     // 실제 Axis가 그려질 영역
     const chartSize = this.units.pos === 'x' ? chartRect.chartWidth : chartRect.chartHeight;
     this.axisPosInfo = {
